requirements.txt
confluent-kafka
apache-airflow-providers-postgres